% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_parse.R
\name{build_re_structures}
\alias{build_re_structures}
\title{Build random-effects structures from parsed bar terms}
\usage{
build_re_structures(bars, mf)
}
\arguments{
\item{bars}{List of bar term info from \code{extract_bars}.}

\item{mf}{Model frame.}
}
\value{
A list with Zt, Lambdat, Lind, theta, lower, flist, cnms, Gp.
}
\description{
Constructs Zt (sparse), Lambdat, Lind, theta, lower, flist, cnms, Gp
from the parsed bar terms and model frame.
}
\keyword{internal}
