% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fglmm_class.R
\docType{class}
\name{fglmmMod-class}
\alias{fglmmMod-class}
\alias{vcov,fglmmMod-method}
\alias{fitted,fglmmMod-method}
\alias{residuals,fglmmMod-method}
\alias{show,fglmmMod-method}
\alias{summary,fglmmMod-method}
\title{S4 class for fastmlm GLMM fit}
\usage{
\S4method{vcov}{fglmmMod}(object, ...)

\S4method{fitted}{fglmmMod}(object, ...)

\S4method{residuals}{fglmmMod}(object, type = "deviance", ...)

\S4method{show}{fglmmMod}(object)

\S4method{summary}{fglmmMod}(object, ...)
}
\arguments{
\item{object}{An \code{fglmmMod} object.}

\item{...}{Additional arguments.}

\item{type}{Character; type of prediction or residuals.}
}
\description{
S4 class for fastmlm GLMM fit
}
\section{Slots}{

\describe{
\item{\code{call}}{The matched call.}

\item{\code{formula}}{The model formula.}

\item{\code{frame}}{The model frame.}

\item{\code{flist}}{List of grouping factors.}

\item{\code{cnms}}{List of random-effect term column names.}

\item{\code{Gp}}{Integer vector of group pointers.}

\item{\code{lower}}{Numeric vector of lower bounds on theta.}

\item{\code{theta}}{Numeric vector of fitted variance parameters.}

\item{\code{beta}}{Numeric vector of fixed-effect coefficients.}

\item{\code{u}}{Numeric vector of spherical random effects.}

\item{\code{eta}}{Numeric vector of linear predictor values.}

\item{\code{mu}}{Numeric vector of fitted means (response scale).}

\item{\code{deviance}}{The deviance at convergence.}

\item{\code{vcov_beta}}{Variance-covariance matrix of fixed effects.}

\item{\code{optinfo}}{List with optimiser convergence info.}

\item{\code{family}}{The GLM family object.}

\item{\code{X}}{Fixed-effects model matrix.}

\item{\code{Zt}}{Sparse transpose of random-effects design matrix.}

\item{\code{Lambdat}}{Sparse relative covariance factor.}

\item{\code{Lind}}{Integer vector mapping theta to Lambdat entries.}
}}

