% Generated by roxygen2: do not edit by hand
\name{fixef}
\alias{fixef}
\title{Extract fixed effects}
\usage{
fixef(object, ...)
}
\arguments{
\item{object}{A fitted model object.}
\item{...}{Additional arguments.}
}
\value{
Named numeric vector of fixed-effect coefficients.
}
\description{
Generic function to extract fixed-effect coefficients from a
fitted mixed model.
}
