% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmlm-package.R
\name{ranef}
\alias{ranef}
\title{Extract random effects}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\value{
A list of data frames, one per grouping factor.
}
\description{
Extract random effects
}
