% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{satterthwaite_df}
\alias{satterthwaite_df}
\title{Compute Satterthwaite degrees of freedom for fixed effects}
\usage{
satterthwaite_df(object)
}
\arguments{
\item{object}{An \code{fmlmMod} object.}
}
\value{
A numeric vector of degrees of freedom, one per fixed effect.
}
\description{
Approximates the denominator degrees of freedom for t-tests of
fixed effects using the Satterthwaite (1946) method.
}
\keyword{internal}
