% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmlm-package.R
\name{VarCorr}
\alias{VarCorr}
\title{Extract variance-covariance of random effects}
\usage{
VarCorr(x, sigma = 1, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{sigma}{Residual standard deviation (used for scaling).}

\item{...}{Additional arguments.}
}
\value{
A list of variance-covariance matrices.
}
\description{
Extract variance-covariance of random effects
}
